Move test for svp fund tx registration to BridgeSupportRegisterBtcTransactionTest class. Minor refactor to reuse bridgeContractAddress
